+++ datatypes/src/commonMain/kotlin/at/asitplus/crypto/datatypes/X509SignatureAlgorithm.kt
@@ -20,8 +20,8 @@ import kotlinx.serialization.encoding.Encoder
  */
 val OID_ECDH_ES = ObjectIdentifier("1.3.132.1.12")
 
-@Serializable(with = CryptoAlgorithmSerializer::class)
-enum class CryptoAlgorithm(override val oid: ObjectIdentifier, val isEc: Boolean = false)
+@Serializable(with = X509SignatureAlgorithmSerializer::class)
+enum class X509SignatureAlgorithm(override val oid: ObjectIdentifier, val isEc: Boolean = false)
     : Asn1Encodable<Asn1Sequence>, Identifiable {
 
     // ECDSA with SHA-size
@@ -95,7 +95,7 @@ enum class CryptoAlgorithm(override val oid: ObjectIdentifier, val isEc: Boolean
         }
     }
 
-    companion object : Asn1Decodable<Asn1Sequence, CryptoAlgorithm> {
+    companion object : Asn1Decodable<Asn1Sequence, X509SignatureAlgorithm> {
 
         @Throws(Asn1OidException::class)
         private fun fromOid(oid: ObjectIdentifier) = runCatching { entries.first { it.oid == oid } }.getOrElse {
@@ -103,7 +103,7 @@ enum class CryptoAlgorithm(override val oid: ObjectIdentifier, val isEc: Boolean
         }
 
         @Throws(Asn1Exception::class)
-        override fun decodeFromTlv(src: Asn1Sequence): CryptoAlgorithm = runRethrowing {
+        override fun decodeFromTlv(src: Asn1Sequence): X509SignatureAlgorithm = runRethrowing {
             when (val oid = (src.nextChild() as Asn1Primitive).readOid()) {
                 ES512.oid, ES384.oid, ES256.oid -> fromOid(oid)
 
@@ -121,7 +121,7 @@ enum class CryptoAlgorithm(override val oid: ObjectIdentifier, val isEc: Boolean
         }
 
         @Throws(Asn1Exception::class)
-        private fun parsePssParams(src: Asn1Sequence): CryptoAlgorithm = runRethrowing {
+        private fun parsePssParams(src: Asn1Sequence): X509SignatureAlgorithm = runRethrowing {
             val seq = src.nextChild() as Asn1Sequence
             val first = (seq.nextChild() as Asn1Tagged).verifyTag(0.toUByte()).single() as Asn1Sequence
 
@@ -157,17 +157,17 @@ enum class CryptoAlgorithm(override val oid: ObjectIdentifier, val isEc: Boolean
     }
 }
 
-object CryptoAlgorithmSerializer : KSerializer<CryptoAlgorithm> {
+object X509SignatureAlgorithmSerializer : KSerializer<X509SignatureAlgorithm> {
 
     override val descriptor: SerialDescriptor =
-        PrimitiveSerialDescriptor("CryptoAlgorithmSerializer", PrimitiveKind.STRING)
+        PrimitiveSerialDescriptor("X509SignatureAlgorithmSerializer", PrimitiveKind.STRING)
 
-    override fun serialize(encoder: Encoder, value: CryptoAlgorithm) {
+    override fun serialize(encoder: Encoder, value: X509SignatureAlgorithm) {
         value.let { encoder.encodeString(it.name) }
     }
 
-    override fun deserialize(decoder: Decoder): CryptoAlgorithm {
+    override fun deserialize(decoder: Decoder): X509SignatureAlgorithm {
         val decoded = decoder.decodeString()
-        return CryptoAlgorithm.entries.first { it.name == decoded }
+        return X509SignatureAlgorithm.entries.first { it.name == decoded }
     }
 }