+++ CHANGELOG.md
@@ -145,3 +145,4 @@
  * Move `curve` from `CryptoAlgorithm` to `JwsAlgorithm`
  * Don't read curve information into the X.509 tea leaves where none exists
    * `CryptoSignature`s in X.509 are now indefinite length
+ * Rename `CryptoAlgorithm` to `X509SignatureAlgorithm` to better describe what it is